Change temperature_measured contract
To validate the data that is passed to the notifictions-service, the
temperature measured contract has to include a field that is forwarded
to the notifications-service as a ground truth.

This commit adds a `measurement_hash` to the temperature_measured event,
which is a hash computed by the load generator. This field can only be
read by entities sharing the key that was used to encrypt the value.