producer: Add log queries
The experiment-producer provides tracing information in a json
structured format into a file. This data can be queried after the fact
for further insights.

This commit adds 2 scripts which are automated queries.
`get-experiments.sh` returns the experiments executed after a period of
time. `get-experiment-temperature.sh` provides the average temperature
values for a given experiment.