Add response error handling and retry logic
When performing a request to a host, different errors might occur
throughout the process of validating the response. This commit adds a
`ResponseError` enum which includes the expected errors throughout this
process.

This commit also refactors some of the code living in `requests.rs` by
breaking up the function that makes the request to a single endpoint
into a request builder and a request validator. Making the actual
request is done by calling the `.send()` method on the RequestBuilder
instance.

Additionally, if the request fails for some kind of ServerError, we
retry the request up until 3 requests have been performed. If
unsuccessful, then we mark the request as failed.